tidy spellcheck improvments based on code review
